social-quote-generator/requirements.txt
# Core dependencies
Pillow>=10.0.0              # Image manipulation
PyYAML>=6.0                 # Configuration parsing
python-frontmatter>=1.0.0   # YAML frontmatter parsing
python-dotenv>=1.0.0        # Environment variable management
tenacity>=8.2.0             # Retry logic
filelock>=3.12.0            # File locking for queue management

# Social media APIs
tweepy>=4.14.0              # Twitter API
instagrapi>=2.0.0           # Instagram API
facebook-sdk>=3.1.0         # Facebook API
requests>=2.31.0            # HTTP requests (for LinkedIn API)

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
types-PyYAML>=6.0.0        # Type stubs for PyYAML
types-Pillow>=10.0.0       # Type stubs for Pillow
